feat: Split process and default process in ProcessValue (#2)
This adds a level of indirection to the value processing.  Now the derived processor
code lives in `process_child_values` which is invoked by the `Processor` trait for
each method.  This means that `process_value` now calls into `Processor::process_X`
which in turn calls into `process_child_values`.

The benefit of this is that state can now be modified before the processor walks
down.  This is useful for trimming operations and the like.