feat: Refactor skip_serialization attribute
Fields now inherit `skip_serialization` to allow modifying the ToValue
behavior of `Object` and `Array`.

In the following example, `null` values are serialized for every level
within `frame.vars`:

```
struct Frame {
    #[metastructure(skip_serialization = "never")]
    vars: Annotated<Object<Value>>
}
```

in this example, an empty `bar.values` would not be serialized
even though the inner `Bar` struct has no annotations. Likewise an empty
`Bar` (having no value for `values`) will not be serialized.

```
struct Foo {
    #[metastructure(skip_serialization = "empty")]
    bar: Annotated<Bar>
}

struct Bar {
    values: Annotated<Array<String>>
}

// Input 1:
// {"bar": {"values": []}}
// Input 2:
// {"bar": {"values": null}}
// Input 3:
// {"bar": {}}
// After:
// {}

```

However, if I were to annotate `values` with
`#[metastructure(skip_serialization = "never")]`, an empty array will
always be serialized while a missing value for `values` will still result in
`bar` not being serialized.

The kickoff for this PR was a bug where a stack frame with `"vars":
{"foo": null}` would end up as a frame without any variables. While this
PR is not really necessary to fix that bug, it still should make it
easier to port interfaces and behavior from Python, as
`skip_serialization` is a property of the attribute there and not of the
type.