ref(general): Change non-string data scrubbing to drop values
For numbers and arrays, we're taking the conservative approach to
completely remove values if their key is blacklisted in the
"sensitiveFields" array. We cannot replace them with "[Filtered]" since
that would change their type.

We keep objects, however, with a blacklisted key. The assumption is that
keys inside these objects will match one of the sensitive fields and
then be stripped.