ref(projects): Jitter project state requests (#277)
Shift each project to a deterministic slot within a grid of expiry intervals (e.g. 5 minutes).

Each project state is still requested immediately when it is seen for the first time. Then it is updated within the next 5 minutes at the end of the current window, and then again at each window boundary. This helps to spread out requests while maintaining a deterministic order.