feat(protocol): Add valid platforms as constant (#294)
Relay currently dismisses all platforms as invalid. Previously, valid platforms was scheduled for removal; however, it turned out that we need to continue validating for analytics purposes.

This PR adds a constant for `VALID_PLATFORMS` instead of the dynamic set in the store config. The set is also exposed to the python layer to be consumed by Sentry. We will remove the list of valid platforms from Sentry once this is released.