ref: Remove never rule type (#471)
IIRC (and take this with a grain of salt), this rule was created to cancel out other rules, back when we stopped at the first matching rule for a value. So if ["my_never_rule", "@anything:remove"] was applied, nothing happened because the first rule matched and we stopped there.

There are two reasons this no longer works:

    We used to have text and container as selectors which were mutually exclusive. Nowadays a field can be matched by multiple selectors at once, and the order in which rules are now applied cannot really be controlled by the user (we just sort selectors alphanumerically)
    we don't stop at the first matching rule because strings can contain both creditcards and IP addresses

Regardless of historical context this is useless right now.