ref: Make processing-relevant filepaths strippable (#469)
In NativeImagePaths we must not strip the file's basename because that would break
processing.

We pop the basename from the end of the string, call process_string and push the
basename again.

The ranges in Meta should still be right as long as we only pop/push from the end of the
string. If we decide that we need to preserve anything other than suffixes all PII
tooltips/annotations are potentially wrong.