feat(pii): Quoted keys in selectors (#485)
Besides the fact that this is probably something the user wants (ability to address keys with whitespace etc in them), we need this to safely roundtrip PII configs generated from legacy datascrubber settings, where all kinds of keys are possible. We would previously emit PII configs that are broken.

Relay-as-store is not affected by this bug, python store is though.