fix(pii): Better selector grammar (#508)
Reintroduction of #501

With #502 I reverted two PRs because I was not sure which of them caused the production impact and we just wanted to get master back to being safe. I believe this may be safe to bring back while we look into the memory leak.

There is potential for increased CPU consumption, in that case there is an obvious way forward with PrecClimber.