ref: Split up benchmarks (#510)
We want to know how much time is spent in initializing the PII processor vs how much time is spent actually walking through the event. Right now nothing interesting is happening when instantiating the PII processor, but in the future there probably will be.

This PR should also make it easier to test matrices of multiple datascrubbing configs vs multiple event payloads if we ever want to go there.