fix(server): Emit outcomes with key_id (#537)
Computes `Scoping` information (organization, project and key) in central places of request handling and passes it through the pipeline. This allows to emit the `key_id` outcome field. Scoping is now computed in the following places:

 - At the beginning of the web request, and then updated when getting project information.
 - At the beginning of envelope handling, and then upated when fetching full project config.
 - For smaller signatures, again in `EventProcessor::enforce_quotas`.

To facilitate this, scoping has been split into item-agnostic `Scoping` and specific `ItemScoping` for rate limiting.