feat(envelope): Introduce an item type for transactions (#555)
Introduces a transaction item type to envelopes and ensures that it is
used consistently.

This made the `event_type` item header obsolete, since it's now no
longer needed to determine the appropriate action for storing an event.

When ingesting envelopes, it is assumed that the event comes in an
`Event` item and transactions come in `transaction` item. To enforce
this, the event_type is overridden, when retrieving items from the
envelope. In normalization, the transaction event type is then left in
place.

The store endpoint needs to be more lenient, however. It checks for the
event type in the JSON payload and puts the event into an appropriate
item.