feat(config): Add a configuration for the shutdown timeout (#561)
The shutdown timeout introduces an unconditional delay after receiving
the shutdown signal. To configure this, a `shutdown_timeout` option is
introduced now that can be overridden in the config.

To achieve this, the `Controller` had to be made configurable. Since it
is a `SystemService`, it needs to retain a `Default` impl, which now
contains no timeout. When starting the service, the default from the
config is filled in.

Along with this goes a small refactor of the `Subscribe` message. Since
the usual pattern is always the same, there is now a shorthand
`Controller::subscribe(Addr)`.