feat(pii): Expose event values as part of generated selectors (#575)
We found that even with the selector autocompletion people sometimes just didn't know what the selectors actually stood for. The best example is $frame.filename vs $frame.abs_path where the only reason those two values exist is a technicalilty (or multiple) inside of Sentry.

Let's expose the actual string values and see what we can do with them in the UI. For the case of abs_path vs filename the difference should be very recognizable.