feat(server): Apply clock-drift correction in proxy Relays (#581)
Exposes the `ClockDriftProcessor` and applies it in non-processing mode.
This allows for proper clock drift correction even when using proxy
Relays that are in remote locations.

This also contains a workaround for the current experimental JavaScript
SDK (versions `5.2` - `5.15`), which relies on clock drift correction
without `sent_at`. In such a case, the `sent_at` value is inferred from
the event's timestamp.