feat(general): Add SpanStatus to span struct (#603)
According to our dev docs at https://develop.sentry.dev/sdk/event-payloads/span/, status should be an explicit field on the span struct. The JS SDK is already setting it, and it is rendered for individual spans in the UI.

Update the Span struct to have a field status. This is typed using SpanStatus, the same as the status on an event. Also updated the tests in the file.