meta: Add unreleased changes to changelog (#605)
Adds unreleased changes to the changelog. For now, changes are grouped
into the following sections:

- "Features": New or enhanced features relevant to users or the public
  behavior of Relay.
- "Bug Fixes": Fixes for the public behavior of Relay or that impact
  usage.
- "Internal": Both changes and bug fixes only relevant to the processing
  mode, trusted Relay operation.

Public behavior is mostly scoped to server features such as endpoints,
compression, rate limits and caching behavior. Data scrubbing is
explicitly included in this feature set. Explicitly excluded are store
normalization and features only for trusted Relays.

This prepares for the `auto` changeset policy in craft.