ref(server): Apply rate limits per item in fast-path (#618)
Introduces `EnvelopeLimiter`, a utility that applies rate limits to envelope items. It implements logic around a `check` function that can either be used to check cached rate limits or Redis. The `check` function is called with the following rules:
 - Once for a single event, if present in the envelope.
 - Once for all attachments comprised, unless the event was rate limited.
 - Once for all sessions comprised.

Items violating the rate limit are removed from the envelope. This follows a set of rules:
 - If the event is removed, all items depending on the event are removed (e.g. attachments).
 - Attachments are not removed if they create events (e.g. minidumps).
 - Sessions are handled separate to all of the above.

This is only used in the fast-path and will be added to normalization in a follow-up.