feat(quotas): Check quota units before applying them (#623)
This checks and removes quotas that Relay cannot handle to avoid invalid
counting. Particularly, this checks that counted quotas only apply to
data categories with the same unit to avoid invalid counters. Units are:

- "count" for all events
- "bytes" for attachments
- "batches" for sessions (for the time)

Counted quotas are all quotas except for those with a limit of `0`. Such
quotas may apply to all categories regardless of their units, since they
are never counted and statically rejected.