ref(server): Build pipeline-like control flow for processing (#635)
Refactors the envelope and event processing code by splitting the main
processing function into mostly independent pipeline steps that all
operate on a state. The state is mutated by each function, which can be
thought of as passing it from one pipeline step to the next.

From the top-level process function, it becomes immediately clear, which
actions are applied with processing and which are skipped. The stages
only required for processing are conditionally compiled.

There is no change in Relay's behavior. However, there are a few changes
internally:

- The flow for rate limiting is different. The function `enforce_quotas`
  is called on every envelope in the pipeline. This prepares the
  pipeline for rate limiting of items other than events. At the moment,
  it skips all envelopes without events and still assumes
  `DataCategory::Error` for all events.
- Clock drift correction now runs unconditionally and has been removed
  from the store normalize processor.
- Processing mode errors if no event payload can be extracted from items
  that should create an event. This includes minidumps with their
  placeholders.
- Rate limits are still sent to the project for caching. However, these
  messages have moved from the global error handler into the processing
  and upstream branches. This means that cached rate limits no longer
  update the project.