feat(server): Emit flags for rate limited attachments (#640)
By default, rate limited attachments are immediately removed from
Envelopes. For processing, native crash reports still need to be
retained. These attachments are marked with the "rate_limited" header,
which signals to the processing pipeline that the attachment should not
be persisted after processing.

Since this requires changes in Sentry, this flag is only added when a
hidden config variable is set. The config will be removed in a future
release of Sentry.