fix(server): Never respond with 429 for minidumps (#646)
Many third-party minidump clients retry requests that are not
successful, including `429 Too Many Requests`. However, such requests
should never be retried. For this reason, always respond with `200 OK`
despite rate limits.

This is limited to the `/minidump/` and `/unreal/` endpoints, since
Sentry clients as well as community SDKs are expected to obey the rate
limit header.