fix(server): Do not 429 for event-only attachments (#647)
Envelopes containing minidumps lead to status code
`429 Too May Requests` if attachment storage is disabled on the server.
This is undesirable, since it is perfectly legal to send standalone
minidumps for processing purposes.

With this patch, rate limits for attachments are only recorded if at
least one attachment is _plain_, which means it does not also create an
event and is required for processing. This creates correct perceived
behavior of the endpoint, by only responding with a rate limit if the
Envelope contains a custom attachment.