feat(py): Expose DataCategory via C-ABI (#651)
Exposes `DataCategory` via the C-ABI to Python for consumption in Sentry
and Snuba. Along with this, also exposes the `SpanStatus`, still via the
old dictionary. In a future change, this could be made an enum, too.

To expose the constants, they have been moved to `relay-cabi`. This
creates a central place for constants that are exposed from Relay to
other services in Sentry.

Along with this, the deprecated serialization of span status
`"unknown_error"` is removed. By now, all downstream consumers have been
updated.