ref(server): Add a header for rate limited items (#660)
Separates the logic for summarizing Envelope contents into a separate
helper. As part of this, the new helper skips items that have been rate
limited before. Otherwise, this would invoke the rate limiter with
redundant quantities.

Additionally, the rate limited header is now typed out in item headers.
However, this header is never serialized or deserialized, so that
clients cannot set it. For Relays, this means that each Relay can
independently determine whether items need to be rate limited, rather
than trusting the downstream.