Do not raise exceptions in teardown (#663)
Any exception in teardown is an error, not a test failure.  Instead
modify the runtest protocol call step to perform the checks that the
fixture wants to do at the end of the tests.

The marker system is a great way to transfer state on the test item
since it uses a plublic API to both add the information as well as
access it.  The marker itself is not specifically tied to this
fixture.

This also modifies the default run to include a report of failed
tests.