feat: Preserve full Span.Description
SDKs store potentially lenghty information in the description that is
only useful when complete. In particular, the DB integration in Python
sets the description to a SQL query. Users looking at slow spans want to
be able to read the SQL query as part of understanding why it was slow.

Truncation after 1024 bytes means often times not even the SELECT part
of the query is readable.

While we may consider changing SDKs to send the full query in a
different field and leverage that in the Sentry UI (e.g. syntax
highlighting), the expedient solution to fix the current UX is to remove
the truncation in Relay.

We could alternatively increase the limit, e.g.

  #[metastructure(max_chars = "message")]

would allow for 8192 bytes, 8x more than 'summary' at 1024 bytes,
however, that becomes just yet another arbitrary truncation point, way
below the limit for total payload size.

Since too large payloads would have been discarded before they hit Relay
anyway, it seems fair to drop the truncation for now and observe.

We can safely revert this change later if necessary.