feat: Preserve full Span.Description (#674)
SDKs store potentially lenghty information in the description that is
only useful when complete. In particular, the DB integration in Python
sets the description to a SQL query. Users looking at slow spans want to
be able to read the SQL query as part of understanding why it was slow.