doc(schema): Export canonical event schema as JSON Schema (#558)
This introduces schemars as dependency and adds a new subcommand to relay that generates the schema, which we then use to publish the schema as release artifact.

Alex Hofsteede attempted to document the event schema using JSON schema too, but sort of got stuck halfway through. I don't know if it was lack of time or something else. I believe the difference is that we only attempt to document the canonical version after store normalization while Alex tried to validate every legacy format as well. The only exception we make there is that tags/headers/cookies are also allowed as objects even though the canonical form is array of arrays.


This schema file's target audience is mostly SDK developers, but maybe Snuba ppl can profit from it too.

Note: We cannot profit from this within the Python SDK as the only tooling for using JSON schema in type annotations is https://pypi.org/project/jsonschema-typed which users would have to install in their devenv to see any effect. What I actually wanted was code generation from JSON schema, not a runtime plugin for the type checker. I think for Snuba it would be fine but it's still inconvenient to set up and the error messages will be very poor for sure.

I am fairly certain that this will be a good data source for auto-generated documentation at the very least. We should merge this before it gets out of sync with master even more, later we can make adjustments to make the autogenerated docs better.