feat(server): Support chunked formdata for Crashpad (#721)
Google Crashpad has a limitation on the length of values for custom
attributes. This makes it impossible to send scope information and event
payloads using the default Crashpad uploader. In the past, our SDKs had
to disable automatic uploading, and implement their own logic, which
comes with a set of disadvantages. In the future, Crashpad will support
custom attachments on all platforms, but until that is available, we
need a workaround on our end.

An upcoming version of Electron will send long crashpad annotations in
chunks to work around this limitation. When a long value is detected,
the annotation is duplicated with a postfix of __n, where n is the
1-based index of the chunk.

This PR joins these chunks, assuming they come in random order, and then
parses JSON to extract the payload. An event id is not allowed in these
payloads, as it cannot be parsed efficiently in the endpoint. Instead,
Relay generates a random event id and always overrides the one in the
payload.