fix(server): Send envelopes to the envelope endpoint (#746)
Envelopes were initially introduced on the `/store/` endpoints, but have
since moved to `/envelope/`. This is now the primary endpoint for
ingesting Envelopes, and sending Envelopes to the old endpoint is
soft-deprecated.

Since the store endpoint has different size limits on Sentry SaaS, Relay
has to send envelopes to the correct endpoint to avoid spurious 413
responses.