doc: Use cargo doc for github pages (#755)
Instead of custom documentation, build `cargo doc` for GitHub pages. The
root crate documentation then links to Sentry docs and, in future,
develop. Architecture documentation will be split:

- develop will contain high-level architecture, mostly around the
  integration of the Relay service into devservices, and interfacing
  with systems like Kafka or Redis.
- crates docs will contain specifics around actors, messages, and the
  control flow of an event.

This PR removes all docs sources, Python requirements, and the Makefile
targets. In a follow-up, we will switch to building docs on pushes to
master, instead of releases.