feat(config): Add an option for upstream content encoding (#771)
Adds the `http.encoding` configuration option to disable or change
compression of upstream store requests. It is planned that this is
eventually applied to all upstream requests, hence the generic naming.

By default, compression should be enabled by all Relays. Sentry applies
more aggressive limits on the content-length of raw requests than after
decompression. However, in the current state, compression can become a
bottleneck in high throughput scenarios, in which case it may be more
beneficial to disable compression.