fix(server): Disable generic network error handling (#772)
Network error handling relies on authentication, which is not available
in proxy and static mode. Additionally, it fails health checks, which
can cause downstream clients to stop sending data to Relay.

As a temporary measure, this PR disables network error handling outside
of authentication until a dedicated mechanism is implemented.