fix(setup): Log when reporting to Sentry is disabled (#779)
The logging integration of Sentry SDK 0.19 and 0.20 supresses log
messages when the client is disabled, for instance if the DSN is set to
None. Since this is the default configuration of Relay, logs no longer
print to standard error by default.

This PR skips initializing the Sentry SDK if error reporting is disabled
and instead registers the logger directly.