Avoid the forbidden transmute word
This does the same magic "please interpret these bits over here as the
same type but with a different lifetime" as the transmute.  But with
the same type being forced by using more words and a custom function
instead of a turbofish.  Thus avoiding the word which shall not be
used.