ref(common): Make the DSN public key Copy (#795)
Creates a type for DSN public keys that implements copy. Every public
key is a 32-character hexadecimal string. Storage requirements could be
halved by parsing the hexadecimal contents of the key, however this is
currently not enforced anywhere in Sentry. For this reason, we're taking
the conservative approach using 32 bytes.

The type is called `ProjectKey` to distinguish it from Relay's own
public key.