fix(server): Do not send None relay action (#796)
This fixes a bug where every Relay error response now included a `relay`
action field, even if the action was None. Instead, to save space and
not confuse users reading the error responses during debugging, this
action should only be serialized if it is not `None`.

Also, this folds the `None` and `Unknown` cases together, since there is
no semantic difference.