ref: Avoid the encoding crate by using WStr and Utf16Error (#792)
This re-models the Utf16Error after std::str::Utf8Error which allows
it to provide enough information to write a decoder.  This means we no
longer need a decoder and can write our segmentation iterator a little
simpler.

The complexity of this is a lot less than using the RawDecoder API
plus there is fewer duplication in the branches, so this is more
maintainable.  Downside is that this has had less production scrutiny.

Co-authored-by: Jan Michael Auer <mail@jauer.org>
Co-authored-by: Arpad Borsos <arpad.borsos@sentry.io>