fix(server): Rate limit outcomes emited only for events on "fast-path" (#809)
Rate limited outcomes are generated on the normal "slow path" but since
rate limits are cached, there is a fast path that checks the cache and
fails fast if know we are above the rate limit. Here we were also
generating invalid outcomes

This fixes outcome generation on the "fast-path" i.e. when we know from
the cache that we are rate limited and we don't try to process the
envelope.