ref(envelopes): Validate userfeedback
This is modelled after session validation.

While doing that I noticed that some of the Sentry tags we set are
useful for both codepaths, so I moved the with_scope outside, which is
more expensive due to allocating new strings 100% of the time, but
cleaner.  Ideally we would be able to assign borrowed data to the scope
at some point to actually make this fast.

I also wanted to test this within Rust, so I changed some of the
constructors to be easier to use from tests.