fix(server): Add a client SDK interface for unreal crashes (#828)
Cleans up unreal testing code and adds a client SDK interface named
`unreal.crashreporter` to the event payload when an unreal context was
detected.

This was omitted in #376 originally, since the unreal crash reporter is
not an SDK. However, the SDK interface is used for analytics. Note that
the original value of this was `sentry.unreal.crashreporter`.