feat(pii): Make $error.value pii=true
There is not really a good reason as to why one would have to jump
through a bunch of hoops to scrub this field. It is mostly equivalent to
`$message` in what it contains (to the point people confuse one with
another) and impacts grouping in similar ways.

The kickoff for this PR was when a customer attempted to configure
`$error.**` and failed.

The original concern with making this pii=true was that it would impact
grouping, but if anything, removing PII/variable data improves grouping.
This is more of a concern with the Default Scrubbers where the entire
value may be removed, which is why the field has been put on the
ignore-list for those.