feat(protocol): Add protocol support for WASM (#852)
This extends the protocol for WASM. In the frame you can now provide
`addr_mode` to override the behavior of an address. For instance one can
set this to `rel:index` to set it relative to the index of a debug
image. This is all handled in the background in sentry.