ref(server): Simplify capture mode implementation (#858)
Removes the `RwLock` around the map of captured envelopes and instead
uses an `ActorFuture` for synchronization. This removes the need to
place the map into an `Arc` and clone it across the future.

Since capture mode no longer only records events but entire envelopes,
the associated types are now called `CapturedEnvelope`.