fix(user-report): Make all fields but event-id optional (#886)
the exact schema has been really ill-defined, and it appears most SDK devs expect those fields to be optional now. We already store empty strings for all of those fields, so let's just make them all optional.

The "schema" for user report is "defined" is as the set of possible keyword-arguments to the Python model itself. Anything goes that the model constructor accepts. Explicit null was never really supported. Also changing to Option means that we would have to coerce to empty string on serialization precisely to avoid running UserReport(email=None).save() which would violate NotNullConstraint of the database.