feat: Enable use of multiple value types per field (#882)
Add ability to assign multiple value types to a protocol type. For example (part of the patch):

#[metastructure(value_type = "Message", value_type = "String")]

This allows us to fix a long-standing issue with logentry.formatted,
which is supposed to be addressible by $string as well as
$logentry.formatted as well as $message.

It also makes a bit of code in minidump attachment processing simpler,
and removes an obstacle in introducing valuetypes in a
backwards-compatible manner.

For future consideration we may change the behavior of the derive on
newtypes to append to the list of valuetypes rather than replace them,
such that the actual syntactic change in logentry.rs is not necessary
(as the Message wraps a String already). Similarly one could make all
structs-with-fields addressible under $object.