fix(server): Do not stream with chunked transfer encoding (#890)
The reqwest client implemenation creates a stream to compress request payloads
on the fly when sending them to the upstream. Unsurprisingly, this switches the
client into "chunked" transfer encoding mode, since it no longer knows the
content length ahead of time.

While this is not an issue for store requests, the Sentry web server does not
understand chunked transfer encoding and rejects such requests. This was not
caught in integration tests, since those build on flask which supports chunked
transfer encoding by default.

This PR reverts request compression to an earlier draft, which compresses the
entire body eagerly.