feat(loader): Add JS loader to relay
This embeds the loader into relay directly.  This requires a change in
Sentry to pass the necessary information relay: https://github.com/getsentry/sentry/pull/23021

Note that even if the Sentry change is not merged the loader gets served
up in some capacy though no fallback is provided for the CDN URL.

A preferred solution is likely to also have sentry serve up a locaiton
for the loader JS file instead of compiling it into relay in which case
a relay does not have not be updated to get the latest JS loader.