ref: Remove Compatibility mode for Session Aggregates
The compatibility mode was added in https://github.com/getsentry/relay/pull/815 because it was waiting on proper support in snuba.
That downstream support was added recently in https://github.com/getsentry/snuba/pull/1492, and compatibility mode was switched off in production via a configuration change.
Also, the snuba change was included in the 21.1 CalVer release, so it is fine to remove compatibility mode and ship this feature in full fidelity in 21.2.