feat: Add a .no-cache option to event sending (#911)
This pull request adds a `no-cache` flag that forces the internal
project cache to refresh before processing the envelope. The flag is
encoded into the DSN or public key so that it works with any SDK and
third-party (including minidumps).

For external Relays including PoPs, this has the potential to
significantly slow down the end-to-end ingestion timing, but makes it
possible to debug issues better by forcing a clean cache before sending.
This is useful for iterating on grouping or filter settings. Request
handling is not impacted by this change, so SDK event submission should
not be impacted.

In general, DSN public keys can now specify optional flags. Flags are
appended to the public key in dotted syntax:

    https://public-key.flag1.flag2@example.com/1234

The flag for cache bypassing is `no-cache`.