fix(ci): Use loopback ip for advertised listeners (#922)
The new Docker version does not proxy the ipv6 loopback by default
anymore. The GitHub actions virtual environment is adding an alias for
localhost pointing to the ipv6 address which causes librdkafka to
resolve localhost into two addresses and round-robin connections. With
the recent update of docker in Ubuntu repositories, this mechanism is
now broken (for the ipv6 address).

Using 127.0.0.1 in advertised listeners will circumvent the IP
resolution.