fix: Ignore false-positive lints from Clippy 1.50 (#926)
The `clippy::unnecessary_wraps` lint also warns when the call site cannot be
modified. Since this is the case for `serde(default)` callbacks, we need to
ignore the lint.

The `clippy::wrong_self_convention` lint is actually correct. However, since
this is a major refactoring, we will take this in a follow-up.