feat(protocol): Add snapshot to the stack trace interface (#927)
Indicates that this stack trace is a snapshot triggered by an external
signal.

If this field is `false`, then the stack trace points to the code that
caused this stack trace to be created. This can be the location of a
raised exception, as well as an exception or signal handler.

If this field is `true`, then the stack trace was captured as part of
creating an unrelated event. For example, a thread other than the
crashing thread, or a stack trace computed as a result of an external
kill signal.