fix(pii): Make and/or work correctly with pii=maybe (#932)
This fixes a bug where OR-ing a specific selector with ** would make it non-specific. This should not be the case universally, the selector should continue to be specific if the specific sub-selector matches. Example:

** || culprit || $string || $frame.module || $frame.abs_path

this should match against culprit, but doesn't because of the disjunction with **.

Right now the only workaround is to create multiple rules for **, culprit, ... each

successor to #726