feat(stats): Add data category to TrackOutcome (#931)
The TrackOutcome struct needs a DataCategory to indicate whether the
outcome is for an error, transaction, or attachment. Relay has the
notion of a DataCategory (and exports it in the sentry-relay Python
code) but doesn't apply it to outcomes.

Not every event emits an outcome from within the Relay component. The
immediate code changes are geared more toward transactions than
attachments. Transactions need little to no more than the category to
distinguish them, but attachments require more changes.

Co-authored-by: Jan Michael Auer <mail@jauer.org>