fix(quotas): Use whole seconds for rate limits (#946)
Relay over-accepts events in the last second of each rate limiting
window. This shows by a RetryAfter(0) returned from the Redis rate
limiting function, which immediately counts as expired.

For each rate limited item, the rate limiter takes the current timestamp
and places it in a window. The RetryAfter value is the duration to the
end of the window, rounded up to whole seconds. The code assumed that
UnixTimestamp cannot contain a sub-second part, and thus computed:
`(expiry - timestamp).as_secs()`. Within the last second of the window,
the difference is less than one second, which is rounded to 0.

To avoid similar issues, this patch removes the sub-second from
UnixTimestamp and makes it whole seconds.