fix(common): Use zero-copy deserialization when FromStr is implemented (#950)
We sometimes implement both FromStr and serde::Deserialize. In cases
where we share the implementation, we provide an explicit definition of
FromStr and use macros to implement serde::Deserialize.

There were two implementations of this macro:
1. relay_common::impl_str_serialization
2. relay_general::impl_str_serde

The implementation in relay_general allows to implement just
deserialization or serialization, but had a bug: It is not zero-copy.
Instead, it always allocates a String, only to then parse it.

In this change, we deduplicate the macro and ensure that there are no
allocations when deserializing.